backend/models/mood_model.py
import base64
import io
import random
import numpy as np

def analyze_image(base64_image):
    """
    Analyze a base64-encoded image for emotions.
    Tries to use DeepFace for real detection, falls back to FER,
    then falls back to mock data.
    """
    try:
        from deepface import DeepFace
        from PIL import Image

        # Decode base64 to image
        # Handle data URI prefix if present
        if ',' in base64_image:
            base64_image = base64_image.split(',')[1]

        img_bytes = base64.b64decode(base64_image)
        img = Image.open(io.BytesIO(img_bytes))
        img_array = np.array(img)

        # Run DeepFace emotion analysis
        results = DeepFace.analyze(img_array, actions=['emotion'], enforce_detection=False)

        if isinstance(results, list):
            results = results[0]

        emotions = results.get('emotion', {})
        primary_emotion = results.get('dominant_emotion', 'Neutral').capitalize()

        # Normalize to percentages (DeepFace already returns percentages)
        # Convert numpy float32 to native Python float for JSON serialization
        confidences = {}
        for emotion, score in emotions.items():
            confidences[emotion.capitalize()] = round(float(score), 1)

        return {
            "primary_emotion": primary_emotion,
            "confidences": confidences
        }

    except ImportError:
        pass
    except Exception as e:
        print(f"DeepFace error: {e}")

    # Fallback: try FER
    try:
        from fer import FER
        from PIL import Image

        if ',' in base64_image:
            base64_image = base64_image.split(',')[1]

        img_bytes = base64.b64decode(base64_image)
        img = Image.open(io.BytesIO(img_bytes))
        img_array = np.array(img)

        detector = FER(mtcnn=False)
        result = detector.detect_emotions(img_array)

        if result and len(result) > 0:
            emotions = result[0]['emotions']
            primary_emotion = max(emotions, key=emotions.get).capitalize()
            confidences = {k.capitalize(): round(v * 100, 1) for k, v in emotions.items()}
            return {
                "primary_emotion": primary_emotion,
                "confidences": confidences
            }
    except ImportError:
        pass
    except Exception as e:
        print(f"FER error: {e}")

    # Final fallback: mock but based on random weighted distribution
    emotions = ['Happy', 'Sad', 'Angry', 'Surprised', 'Neutral', 'Fear', 'Disgust']
    primary = random.choice(emotions)
    raw = {e: random.uniform(1, 15) for e in emotions}
    raw[primary] = random.uniform(50, 95)
    total = sum(raw.values())
    confidences = {e: round((v / total) * 100, 1) for e, v in raw.items()}

    return {
        "primary_emotion": primary,
        "confidences": confidences
    }
