backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
gunicorn==22.0.0
pymongo==4.16.0
python-dotenv==1.0.1
numpy
Pillow
deepface
tf-keras
